services/requirements.txt
# Dependencies for Lambdas using CodeUri: services (PaymentSettle, StorageUpload, etc.).
# SAM installs from this file when building the services artifact.
boto3
eth-account
web3
strawberry-graphql>=0.252.0
mangum>=0.17.0
